+++ xphyle/urls.py
@@ -10,13 +10,15 @@ from urllib.parse import urlparse
 from urllib.request import urlopen, Request
 from xphyle.types import Url, Range, Any, cast
 
+
 # URLs
 
-def parse_url(url_string: str) -> Url:
+
+def parse_url(url_string: str) -> [Url, None]:
     """Attempts to parse a URL.
     
     Args:
-        s: String to test.
+        url_string: String to test.
     
     Returns:
         A 6-tuple, as described in ``urlparse``, or  None if the URL cannot be
@@ -29,13 +31,14 @@ def parse_url(url_string: str) -> Url:
         return None
     return url
 
+
 def open_url(
         url_string: str, byte_range: Range = None, headers: dict = None,
         **kwargs) -> Any:
     """Open a URL for reading.
     
     Args:
-        url: A valid url string.
+        url_string: A valid url string.
         byte_range: Range of bytes to read (start, stop).
         headers: dict of request headers.
         kwargs: Additional arguments to pass to `urlopen`.
@@ -63,11 +66,11 @@ def open_url(
             return io.BufferedReader(cast(HTTPResponse, response))
         else:
             return response
-        return response
     except (URLError, ValueError):
         return None
 
-def get_url_mime_type(response: Any) -> str:
+
+def get_url_mime_type(response: Any) -> [str, None]:
     """If a response object has HTTP-like headers, extract the MIME type
     from the Content-Type header.
     
@@ -81,9 +84,11 @@ def get_url_mime_type(response: Any) -> str:
         return response.headers['Content-Type']
     return None
 
+
 CONTENT_DISPOSITION_RE = re.compile('filename=([^;]+)')
 
-def get_url_file_name(response: Any, parsed_url: Url = None) -> str:
+
+def get_url_file_name(response: Any, parsed_url: Url = None) -> [str, None]:
     """If a response object has HTTP-like headers, extract the filename
     from the Content-Disposition header.
     
@@ -105,6 +110,6 @@ def get_url_file_name(response: Any, parsed_url: Url = None) -> str:
     if parsed_url and hasattr(parsed_url, 'path'):
         # ISSUE: ParseResult has named attributes that mypy does not
         # yet recognize
-        #return parsed_url.path
+        # return parsed_url.path
         return parsed_url[2]
     return None